+++ ArcdpsLogManager/Logs/Updates/LogDataUpdater.cs
@@ -188,11 +188,15 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Updates
 			new LogUpdate(log => log.ParsingVersion < new Version(1, 11, 0, 1)
 			                     && log.Encounter == Encounter.CosmicObservatory,
 				"Fix success detection for Cosmic Observatory"),
-			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
-			// unless the version changes more significantly, in that case it can be reset to 0.
 			new LogUpdate(log => log.ParsingVersion < new Version(1, 11, 0, 2)
 			                     && log.Encounter == Encounter.TempleOfFebe,
 				"Add support for Temple of Febe Legendary CM"),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 11, 1, 0)
+			                     && log.Encounter == Encounter.TempleOfFebe
+			                     && log.GameBuild >= GameBuilds.TempleOfFebeHealthFix,
+				"Fixed NM detection for Temple of Febe"),
+			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
+			// unless the version changes more significantly, in that case it can be reset to 0.
 		};
 
 		public IEnumerable<LogUpdateList> GetUpdates(IEnumerable<LogData> logs)